Move SSO methods over into Client as AuthenticationService will go away very soon
